Updated the greedy implementation of the MapRETEQueue to hold a HashedCircularPriorityQueue instead of a CircularPriorityQueue.
Created a HashedCircularPriorityQueue (currently not size limiting), which stores items in queues dedicated to the hash of the items' shared variables (used as the key in a hash map).  Although this is the intended purpose, the HashedCircularPriorityQueue is fully generic'ed.

In order to fully describe the two classes, where they are similar and where they are different, a hierarchy of TimedX interfaces has been created (TimedCollection -> TimedQueue and TimedMap) to encompass the prune() method, and the extensions to the Queue and Map interfaces to allow specification of an item's temporal data (timestamp and delay).