Migrated some Node creation from the deprecated static methods in Node to those in NodeFactory.
Migrated some RuntimeExceptions to UnsupportedOperationExceptions (more accurate).

Made a debug message in the RIFTripleConsequence more illuminating.

Added a debug message to the JoinStream and KestrelStream for every emission.

Added a generalisation to the Log4jOperation, to detect any lsbench query and report the result along with the query number.

Changed the CSGreedyPlannerMode to default to a window duration of 1 minute (as before), but with a granularity (step) of SECONDS rather than MINUTES