Started debugging issues with the kayo deserialisation:
Because of the holding of NamedNodes in WrappedIVFunctions and WrappedContextISources, NamedNodes were being serialised and deserialised during the build phase.  This is incorrect behaviour, but became more so when the NamedNodes and NamedStreams were edited to be chained together without reference to the over-all graph, as it meant that the entire graph was copied to each function bolt and source spout.

Also extracted the Input during the preparation for deserialisation into its own variable.