Migrated Functions out of providers.
Fixed the use of the Count class during variable anonimisation.

Switched from Node_RuleVariable to Node_Variable in all RIF functions, as RuleVariables are only useful as part of complete Jena rules.

Moved all RIF functions to a consistent package hierarchy.

Added a concept of IConsequence and IPredicate, which provide methods for declaring a functions immediate precursor variable holder in a rule/query plan, for identification purposes.

Added a load of function tests to locate The Bug